# [fix][dingo-executor] Fix optimistic transaction open dingo_constrain…