[fix][dingo-executor] Fix optimistic transaction open dingo_constraint_check_in_place variable using transaction snapshots
show variables like '%constraint_check_in_place%';
+---------------------------------+-------+
| Variable_name                   | Value |
+---------------------------------+-------+
| dingo_constraint_check_in_place | on    |
+---------------------------------+-------+
1 row in set (0.04 sec)

mysql> select * from t1;
+------+------+
| ID   | NAME |
+------+------+
|    1 | xx1  |
|    2 | xx2  |
+------+------+
2 rows in set (0.07 sec)

mysql> begin;
Query OK, 0 rows affected (0.04 sec)

mysql> select * from t1;
+------+------+
| ID   | NAME |
+------+------+
|    1 | xx1  |
|    2 | xx2  |
+------+------+
2 rows in set (0.12 sec)

mysql> insert into t1 values(1, 'xx1');
ERROR 1062 (23000):  Duplicate entry '1' for key 'PRIMARY'
mysql> 
mysql> commit;
Query OK, 0 rows affected (0.02 sec)

mysql>